REFACTOR: 도메인 클래스 재정의
